fix sign-in
